Moved dragContext functionality out of ModView into DragContextManager and ActionPopup
