+++ RoadBuilder/UI/src/mods/ModView/ModView.tsx
@@ -1,7 +1,7 @@
 import { startTransition, useContext, useEffect, useMemo, useRef, useState } from "react";
 import { BottomView } from "../BottomView/BottomView";
 import { SidePanel } from "mods/SidePanel/SidePanel";
-import { DragContext, DragContextData } from "mods/Contexts/DragContext";
+import { DragContext, DragContextData, DragContextManager } from "mods/Contexts/DragContext";
 import { NetSectionItem } from "domain/NetSectionItem";
 import { Number2 } from "cs2/ui";
 import { LaneListItemDrag } from "../Components/LaneListItem/LaneListItem";
@@ -24,14 +24,8 @@ export const ModView = () => {
   const roadBuilderToolMode = useValue(roadBuilderToolMode$);
   let rem = useRem();
   let { translate } = useLocalization();
-
-  let [draggingItem, setDraggingItem] = useState<NetSectionItem | undefined>();
-  let [draggingLane, setDraggingLane] = useState<RoadLane | undefined>();
-  let [mousePosition, setMousePosition] = useState<Number2>({ x: 0, y: 0 });
-  let [mouseReleased, setMouseReleased] = useState<boolean>(false);
-  let [fromDragIndex, setFromDragIndex] = useState<number | undefined>(undefined);
-  let [actionPopupPosition, setActionPopupPosition] = useState<Number2>({ x: 0, y: 0 });
-  let dragItemRef = useRef<HTMLDivElement>(null);
+  let dragCtx = useContext(DragContext);
+  let [actionPopupPosition, setActionPopupPosition] = useState<Number2>({ x: 0, y: 0 });  
   let allNetSections = useValue(allNetSections$);
   let defaultLanePropCtx = useContext(LanePropertiesContext);
   let [lanePropState, setLanePropState] = useState<LanePropertiesContextData>(defaultLanePropCtx);
@@ -64,94 +58,7 @@ export const ModView = () => {
       return record;
     }, {});
     return nStore;
-  }, [allNetSections]);
-
-  let onNetSectionItemChange = (item?: NetSectionItem) => {
-    setDraggingItem(item);
-    setMouseReleased(false);
-  };
-
-  let onMouseMove = (evt: MouseEvent) => {
-    if (draggingItem == undefined) {
-      startTransition(() => {
-        setMousePosition({ x: evt.clientX, y: evt.clientY });
-      });
-    } else {
-      setMousePosition({ x: evt.clientX, y: evt.clientY });
-    }
-  };
-
-  let onMouseClick = (evt: MouseEvent) => {};
-
-  let onMouseDown = (evt: MouseEvent) => {};
-
-  let onMouseRelease = (evt: MouseEvent) => {
-    let isDragging = draggingItem != undefined || draggingLane != undefined;
-    if (evt.button == MouseButtons.Primary && isDragging) {
-      setMouseReleased(true);
-    } else if (evt.button == MouseButtons.Secondary && isDragging) {
-      setDraggingItem(undefined);
-      setDragRoadLane(undefined, undefined);
-    }
-  };
-
-  let setDragRoadLane = (lane?: RoadLane, currentIndex?: number) => {
-    setFromDragIndex(currentIndex);
-    setDraggingLane(lane);
-    setMouseReleased(false);
-  };
-
-  let dragData: DragContextData = useMemo(
-    () => ({
-      onNetSectionItemChange: onNetSectionItemChange,
-      mousePosition: mousePosition,
-      netSectionItem: draggingItem,
-      roadLane: draggingLane,
-      setRoadLane: setDragRoadLane,
-      mouseReleased: mouseReleased,
-      dragElement: dragItemRef.current,
-      oldIndex: fromDragIndex,
-    }),
-    [mousePosition, draggingItem, draggingLane, mouseReleased, dragItemRef.current, fromDragIndex]
-  );
-
-  useEffect(() => {
-    switch (roadBuilderToolMode) {
-      case RoadBuilderToolModeEnum.ActionSelection:
-        let halfPopupWidth = (rem * 500) / 2;
-        let halfPopupHeight = (rem * 120) / 2;
-        let bodySize = document.body.getBoundingClientRect();
-        let deltaRight = bodySize.width - (mousePosition.x + halfPopupWidth);
-        let deltaLeft = mousePosition.x - halfPopupWidth;
-        let deltaTop = mousePosition.y - halfPopupHeight;
-        let deltaBottom = bodySize.height - (mousePosition.y + halfPopupHeight);
-        let nPos = mousePosition;
-        if (deltaRight < 0 || deltaLeft < 0) {
-          nPos = { ...nPos, x: nPos.x + Math.min(deltaRight, deltaLeft < 0 ? -deltaLeft : 0) };
-        }
-        if (deltaBottom < 0 || deltaTop < 0) {
-          nPos = { ...nPos, y: nPos.y + Math.min(deltaBottom, deltaTop < 0 ? -deltaTop : 0) };
-        }
-        setActionPopupPosition(nPos);
-        break;
-      default:
-        break;
-    }
-  }, [roadBuilderToolMode]);
-
-  useEffect(() => {
-    document.addEventListener("mousemove", onMouseMove);
-    document.addEventListener("mousedown", onMouseDown);
-    document.addEventListener("mouseup", onMouseRelease);
-    document.addEventListener("click", onMouseClick);
-    console.log("Setup event listener");
-    return () => {
-      document.removeEventListener("mousemove", onMouseMove);
-      document.removeEventListener("mousedown", onMouseDown);
-      document.removeEventListener("mouseup", onMouseRelease);
-      document.removeEventListener("click", onMouseClick);
-    };
-  }, [draggingItem, draggingLane]);
+  }, [allNetSections]);  
 
   let content: JSX.Element | null = null;
   switch (roadBuilderToolMode) {
@@ -164,7 +71,6 @@ export const ModView = () => {
               {translate("Prompt[PickerHint]", "Click On A Road")}
             </span>
           </div>
-          <LaneListItemDrag ref={dragItemRef} />
           <SidePanel />
         </>
       );
@@ -172,7 +78,7 @@ export const ModView = () => {
     case RoadBuilderToolModeEnum.ActionSelection:
       content = (
         <>
-          <ActionPopup popupPosition={actionPopupPosition} />
+          <ActionPopup />
         </>
       );
       break;
@@ -183,8 +89,7 @@ export const ModView = () => {
         <>
           <SidePanel />
           <BottomView />
-          <RoadPropertiesPanel />
-          <LaneListItemDrag ref={dragItemRef} />          
+          <RoadPropertiesPanel />                    
         </>
       );
       break;
@@ -192,12 +97,12 @@ export const ModView = () => {
       return <></>;
   }
   return (
-    <DragContext.Provider value={dragData}>
+    <DragContextManager>    
       <NetSectionsStoreContext.Provider value={nStore}>
         <LanePropertiesContext.Provider value={lanePropCtx}>
           <div className={styles.view}>{content}</div>
         </LanePropertiesContext.Provider>
-      </NetSectionsStoreContext.Provider>
-    </DragContext.Provider>
+      </NetSectionsStoreContext.Provider>    
+    </DragContextManager>
   );
 };
